master: initial support for message signing
Operates (almost) exactly the same way the AWS4 scheme does.

Clients are expected to put a "Authorization" header in the AMQP
basic.properties struct.

HMAC-{NULL,SHA{224,256,384,512}} algorithms are supported for signing.
HMAC-SHA256 is the default. Any outgoing messages are signed with
the default algorithm. Incoming messages will be verified using the
algorithm specified in the "Authorization" header.

There is only very rudimentary support for a cryptographic nonce.
A nonce is provided in the Authorization header, but is currently not
verified. This is intended for future work.

The access key is the string representation of the agent's UUID with
all the inner dashes ('-') removed.

The secret key is hardcoded to "abc123" for all agents. This will be
fixed in a future patch once the relevant infrastructure has been added
to Nimrod.

Example:
  NIM1-HMAC-SHA256
   Credential=00000000000000000000000000000000/20200519T035924Z/0/nimrod,
   SignedProperties=app-id;content-encoding;content-type;delivery-mode;message-id;timestamp;type;user-id,
   SignedHeaders=user-agent;x-nimrodg-sent-at,
   Signature=ff261c65f718620d16dfecc2a5bf6d998e25951347bf845852599c7f5e7cadc3