portal-backend: run new setup code every time
THIS IS A PRIME CANDIDATE FOR A FUTURE REVERT.

The new setup code is idempotent, so it's safer to run it as part
of creating the NimrodAPI instance. However, it (probably) slightly slower
as it has to check migrations.

In the event that we ever get enough for users, this should be reverted
and things done properly.