master: mark stale agents as expired, don't expire them manually
If an orphaned agent was marked as stale, it was adopted by the Orphanage and
then had Master#doExpire() invoked on it. This had the problem of not removing it
from certain internal data structures.

This changes the adoption code to tell mark the agent as expired via the heartbeating code,
so that its expiration and cleanup will be handled in the same way as normal expiration.