portal-backend: add "rmq init" command
This is mostly useful for use with the RabbitMQ Kubernetes operator.
It puts the initial admin credentials in a secret, which is a right pain
to get when using Terraform. This can run as an init container, and
create the user account in the config file, using the secret as a volume
mount.